refactor: inputs
- A bit less redundancy between local/net/replay
- B/F conversion from L/R is now handled in a single place, allowing the netplay functions to only care about the raw inputs
- AssertInput now works anywhere
- AssertInput can no longer bypass SOCD resolution

+++ src/config.go
@@ -207,7 +207,7 @@ type Config struct {
 	} `ini:"Netplay"`
 	Input struct {
 		ButtonAssist               bool    `ini:"ButtonAssist"`
-		SOCDResolution             int32   `ini:"SOCDResolution"`
+		SOCDResolution             int     `ini:"SOCDResolution"`
 		ControllerStickSensitivity float32 `ini:"ControllerStickSensitivity"`
 		XinputTriggerSensitivity   float32 `ini:"XinputTriggerSensitivity"`
 	} `ini:"Input"`
@@ -322,7 +322,7 @@ func (c *Config) normalize() {
 	c.SetValueUpdate("Sound.WavChannels", Clamp(c.Sound.WavChannels, 1, 256))
 	c.SetValueUpdate("Sound.PauseMasterVolume", int(Clamp(int32(c.Sound.PauseMasterVolume), 0, 100)))
 	c.SetValueUpdate("Sound.MaxBGMVolume", int(Clamp(int32(c.Sound.MaxBGMVolume), 100, 250)))
-	c.SetValueUpdate("Input.SOCDResolution", Clamp(c.Input.SOCDResolution, 0, 4))
+	c.SetValueUpdate("Input.SOCDResolution", int(Clamp(int32(c.Input.SOCDResolution), 0, 4)))
 	switch c.Video.MSAA {
 	case 0, 2, 4, 6, 8, 16, 32:
 	default:
